Move link detection into separate class
