+++ application/src/main/java/org/togetherjava/tjbot/features/tags/TagCommand.java
@@ -18,6 +18,7 @@ import net.dv8tion.jda.api.utils.messages.MessageEditBuilder;
 
 import org.togetherjava.tjbot.features.CommandVisibility;
 import org.togetherjava.tjbot.features.SlashCommandAdapter;
+import org.togetherjava.tjbot.features.utils.LinkDetections;
 import org.togetherjava.tjbot.features.utils.LinkPreview;
 import org.togetherjava.tjbot.features.utils.LinkPreviews;
 import org.togetherjava.tjbot.features.utils.StringDistances;
@@ -93,7 +94,7 @@ public final class TagCommand extends SlashCommandAdapter {
             .map(OptionMapping::getAsUser)
             .map(User::getAsMention);
 
-        List<String> links = LinkPreviews.extractLinks(tagContent)
+        List<String> links = LinkDetections.extractLinks(tagContent, true, true)
             .stream()
             .limit(Message.MAX_EMBED_COUNT - 1L)
             .toList();
